improv. remove warnings

+++ src/main/java/fr/nivcoo/challenges/challenges/challenges/types/internal/ConsumeType.java
@@ -1,5 +1,8 @@
 package fr.nivcoo.challenges.challenges.challenges.types.internal;
 
+import fr.nivcoo.challenges.challenges.Challenge;
+import fr.nivcoo.challenges.challenges.challenges.ChallengeType;
+import fr.nivcoo.challenges.challenges.challenges.Types;
 import org.bukkit.Material;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
@@ -7,35 +10,31 @@ import org.bukkit.event.Listener;
 import org.bukkit.event.player.PlayerItemConsumeEvent;
 import org.bukkit.inventory.ItemStack;
 
-import fr.nivcoo.challenges.challenges.Challenge;
-import fr.nivcoo.challenges.challenges.challenges.ChallengeType;
-import fr.nivcoo.challenges.challenges.challenges.Types;
-
 public class ConsumeType extends ChallengeType implements Listener {
 
-	public ConsumeType() {
-		type = Types.CONSUME;
-	}
+    public ConsumeType() {
+        type = Types.CONSUME;
+    }
 
-	@SuppressWarnings("deprecation")
-	@EventHandler
-	public void onPlayerItemConsumeEvent(PlayerItemConsumeEvent e) {
-		if (!checkRequirements())
-			return;
-		Challenge selectedChallenge = getSeletedChallenge();
-		Player p = e.getPlayer();
+    @SuppressWarnings("deprecation")
+    @EventHandler
+    public void onPlayerItemConsumeEvent(PlayerItemConsumeEvent e) {
+        if (!checkRequirements())
+            return;
+        Challenge selectedChallenge = getSeletedChallenge();
+        Player p = e.getPlayer();
 
-		ItemStack is = e.getItem();
-		if (is == null)
-			return;
+        ItemStack is = e.getItem();
+        if (is == null)
+            return;
 
-		Material type = is.getType();
+        Material type = is.getType();
 
-		boolean allow = selectedChallenge.isInMaterialsRequirement(type, (int) is.getData().getData())
-				&& type.isEdible();
-		if (allow)
-			addScoreToPlayer(p);
+        boolean allow = selectedChallenge.isInMaterialsRequirement(type, is.getData().getData())
+                && type.isEdible();
+        if (allow)
+            addScoreToPlayer(p);
 
-	}
+    }
 
 }
