+++ src/main/java/fr/nivcoo/challenges/challenges/challenges/types/internal/FishingType.java
@@ -1,5 +1,8 @@
 package fr.nivcoo.challenges.challenges.challenges.types.internal;
 
+import fr.nivcoo.challenges.challenges.Challenge;
+import fr.nivcoo.challenges.challenges.challenges.ChallengeType;
+import fr.nivcoo.challenges.challenges.challenges.Types;
 import org.bukkit.entity.Entity;
 import org.bukkit.entity.Item;
 import org.bukkit.entity.Player;
@@ -8,34 +11,30 @@ import org.bukkit.event.Listener;
 import org.bukkit.event.player.PlayerFishEvent;
 import org.bukkit.inventory.ItemStack;
 
-import fr.nivcoo.challenges.challenges.Challenge;
-import fr.nivcoo.challenges.challenges.challenges.ChallengeType;
-import fr.nivcoo.challenges.challenges.challenges.Types;
-
 public class FishingType extends ChallengeType implements Listener {
 
-	public FishingType() {
-		type = Types.FISHING;
-	}
+    public FishingType() {
+        type = Types.FISHING;
+    }
 
-	@SuppressWarnings("deprecation")
-	@EventHandler
-	public void onPlayerFishEvent(PlayerFishEvent e) {
-		if (!checkRequirements())
-			return;
-		Challenge selectedChallenge = getSeletedChallenge();
-		Player p = e.getPlayer();
+    @SuppressWarnings("deprecation")
+    @EventHandler
+    public void onPlayerFishEvent(PlayerFishEvent e) {
+        if (!checkRequirements())
+            return;
+        Challenge selectedChallenge = getSeletedChallenge();
+        Player p = e.getPlayer();
 
-		Entity caught = e.getCaught();
-		PlayerFishEvent.State state = e.getState();
-		if (state.equals(PlayerFishEvent.State.CAUGHT_FISH) && caught instanceof Item) {
-			Item itemCaught = (Item) caught;
-			ItemStack is = itemCaught.getItemStack();
-			boolean allow = selectedChallenge.isInMaterialsRequirement(is.getType(), (int) is.getData().getData());
-			if (allow)
-				addScoreToPlayer(p);
-		}
+        Entity caught = e.getCaught();
+        PlayerFishEvent.State state = e.getState();
+        if (state.equals(PlayerFishEvent.State.CAUGHT_FISH) && caught instanceof Item) {
+            Item itemCaught = (Item) caught;
+            ItemStack is = itemCaught.getItemStack();
+            boolean allow = selectedChallenge.isInMaterialsRequirement(is.getType(), is.getData().getData());
+            if (allow)
+                addScoreToPlayer(p);
+        }
 
-	}
+    }
 
 }
